JIT: Keep locals linked list in Statement::m_treeList
Statements have only one field for the "tree list" even though
bidirectional iteration is supported after nodes have been linked. This
worked fine before the locals tree list existed since the "root node"
was always the last node.

This does not work for the locals tree list since the "root node" is not
part of the list, meaning that we need somewhere else to store the
'last' node. Before this PR the assumption was that the root node was
_never_ part of the locals linked list, so we could use the
gtNext/gtPrev fields of the root node. However, the added test case
shows that in fact it is possible we end up with a top level local.

This PR fixes the problem by allocating a stand-in node in this scenario
and keeping it in Statement::m_treeList.

Fix #81018