JIT: Factor DFS and loop finding
* Factor the general DFS computation and encapsulate its result in a
  `FlowGraphDfsTree` class
* Switch profile synthesis to use the general DFS that takes EH
  successors into account
* Factor the loop finding of profile synthesis and encapsulate its
  result in a `FlowGraphNaturalLoops` class

Some diffs are expected when profile synthesis is enabled, since the
loop identification done here handles EH edges, while the old DFS did
not. For example, for code like

```csharp
internal static bool DeclaresType(this MetadataReader reader, Func<MetadataReader, TypeDefinition, bool> predicate)
{
    foreach (TypeDefinitionHandle handle in reader.TypeDefinitions)
    {
        try
        {
            var typeDef = reader.GetTypeDefinition(handle);
            if (predicate(reader, typeDef))
            {
                return true;
            }
        }
        catch (BadImageFormatException)
        {
        }
    }

    return false;
}
```
we would previously not identify that there are 6 blocks inside the
loop, and that there are 2 backedges to the loop head (one from the try
block, one from the catch block). That's because the handler block would
not appear correctly in the DFS.